# perf: reduce allocations in some formatting and validation functions