Two nice performance / memory improvements. 

<details>
<summary>numericField</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/moov-io/ach
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkNumericFieldShort
BenchmarkNumericFieldShort-16    	97991214	       118.7 ns/op	      24 B/op	       3 allocs/op
BenchmarkNumericFieldLong
BenchmarkNumericFieldLong-16     	338344930	        36.13 ns/op	       8 B/op	       1 allocs/op
PASS
ok  	github.com/moov-io/ach	28.253s
```

```
goos: darwin
goarch: amd64
pkg: github.com/moov-io/ach
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkNumericFieldShort
BenchmarkNumericFieldShort-16    	186805923	        64.67 ns/op	      16 B/op	       2 allocs/op
BenchmarkNumericFieldLong
BenchmarkNumericFieldLong-16     	398885554	        30.35 ns/op	       8 B/op	       1 allocs/op
PASS
ok  	github.com/moov-io/ach	34.381s
```

</details>


<details>
<summary>CalculateCheckDigit</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/moov-io/ach
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkCalculateCheckDigit
BenchmarkCalculateCheckDigit/valid
BenchmarkCalculateCheckDigit/valid-16         	85632897	       139.3 ns/op	      32 B/op	       8 allocs/op
PASS
ok  	github.com/moov-io/ach	12.267s
```

```
goos: darwin
goarch: amd64
pkg: github.com/moov-io/ach
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkCalculateCheckDigit
BenchmarkCalculateCheckDigit/valid
BenchmarkCalculateCheckDigit/valid-16         	437170713	        27.36 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/moov-io/ach	15.385s
```